pyproject.toml
[tool.poetry]
name = "ndn-python-repo"
version = "0.4"
description = "An NDN Repo implementation using Python"
authors = ["Zhaoning Kong <jonnykong@cs.ucla.edu>"]
maintainers = [
    "Xinyu Ma <ma.xinyu.26a@kyoto-u.jp>",
    "Tianyuan Yu <tianyuan@cs.ucla.edu>",
]
license = "Apache-2.0"
readme = "README.rst"
homepage = "https://ndn-python-repo.readthedocs.io"
repository = "https://github.com/UCLA-IRL/ndn-python-repo"
documentation = "https://ndn-python-repo.readthedocs.io"
keywords = ["NDN"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Topic :: Database",
    "Topic :: Internet",
    "Topic :: System :: Networking",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

packages = [{include = "ndn_python_repo"}]
include = [
    { path = "tests", format = "sdist" },
    { path = "ndn_python_repo/ndn-python-repo.service", format = ["sdist", "wheel"] },
    { path = "ndn_python_repo/ndn-python-repo.conf.sample", format = ["sdist", "wheel"] },
]

[tool.poetry.dependencies]
python = "^3.10"
python-ndn = "^0.4.2"
pyyaml = "^6.0"
pendulum = "^3.0.0"
setuptools = "^69.0.0"

# Extra dependencies [dev]
pytest = { version = "^8.0.0", optional = true }
pytest-cov = { version = "^4.1.0", optional = true }
plyvel = { version = "^1.5.0", optional = true }
pymongo = { version = "^4.4.1", optional = true }

# Extra dependencies [docs]
Sphinx = { version = "^7.1.2", optional = true }
sphinx-rtd-theme = { version = "^2.0.0", optional = true }
sphinx-autodoc-typehints = { version = "^1.24.0", optional = true }

[tool.poetry.extras]
docs = ["Sphinx", "sphinx-rtd-theme", "sphinx-autodoc-typehints"]
dev = ["pytest", "pytest-cov", "pymongo"]
leveldb = ["plyvel"]
mongodb = ["pymongo"]

[tool.poetry.scripts]
ndn-python-repo = "ndn_python_repo.cmd.main:main"
ndn-python-repo-install = "ndn_python_repo.cmd.install:main"
ndn-python-repo-port = "ndn_python_repo.cmd.port:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
